Move test DNS handler to the test package
