+++ pkg/tools/dnsutils/dnsconfigs/tcp_dns_server_test.go
@@ -23,6 +23,8 @@ import (
 	"context"
 	"fmt"
 	"net"
+	"net/url"
+	"strings"
 	"testing"
 	"time"
 
@@ -35,7 +37,6 @@ import (
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils"
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/dnsconfigs"
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/fanout"
-	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/health"
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/next"
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/noloop"
 	"github.com/networkservicemesh/sdk/pkg/tools/dnsutils/searches"
@@ -103,7 +104,7 @@ type udpHandler struct {
 func (h *udpHandler) ServeDNS(rw dns.ResponseWriter, m *dns.Msg) {
 	name := m.Question[0].Name
 
-	if name == health.ToWildcardPath(healthCheckHost) {
+	if name == ToWildcardPath(healthCheckHost) {
 		resp := prepareIPResponse(name, resolvedIP, m)
 		if err := rw.WriteMsg(resp); err != nil {
 			dns.HandleFailed(rw, m)
@@ -179,14 +180,14 @@ func Test_TCPDNSServerTimeout(t *testing.T) {
 		},
 	})
 
-	healthCheckParams := &health.Params{
+	healthCheckParams := &HealthCheckParams{
 		DNSServerIP: proxyAddr,
 		HealthHost:  healthCheckHost,
 		Scheme:      "udp",
 	}
 
 	clientDNSHandler := next.NewDNSHandler(
-		health.NewDNSHandler(*healthCheckParams, []health.Option{}...),
+		NewTestDNSHandler(*healthCheckParams, []TestDNSOption{}...),
 		dnsconfigs.NewDNSHandler(dnsConfigsMap),
 		searches.NewDNSHandler(),
 		noloop.NewDNSHandler(),
@@ -215,4 +216,138 @@ func Test_TCPDNSServerTimeout(t *testing.T) {
 
 	err = proxy.shutdown()
 	require.NoError(t, err)
+}
+
+// HealthCheckParams required parameters for the health check handler
+type HealthCheckParams struct {
+	// DNSServerIP provides the IP address of the DNS server that will be checked for connectivity
+	DNSServerIP string
+	// HealthHost value used to check the DNS server
+	// DNS server should resolve it correctly
+	// Make sure that you properly configure server first before using this handler
+	// Note: All other requests that do not match the HealthPath will be forwarded without any changes or actions
+	HealthHost string
+	// Scheme "tcp" or "udp" connection type
+	Scheme string
+}
+
+type dnsHealthCheckHandler struct {
+	dnsServerIP string
+	healthHost  string
+	scheme      string
+	dnsPort     uint16
+}
+
+func (h *dnsHealthCheckHandler) ServeDNS(ctx context.Context, rw dns.ResponseWriter, msg *dns.Msg) {
+	name := msg.Question[0].Name
+
+	if name != h.healthHost {
+		next.Handler(ctx).ServeDNS(ctx, rw, msg)
+		return
+	}
+
+	newMsg := msg.Copy()
+	newMsg.Question[0].Name = dns.Fqdn(name)
+	dnsIP := url.URL{Scheme: h.scheme, Host: h.dnsServerIP}
+
+	deadline, _ := ctx.Deadline()
+	timeout := time.Until(deadline)
+
+	var responseCh = make(chan *dns.Msg)
+
+	go func(u *url.URL, msg *dns.Msg) {
+		var client = dns.Client{
+			Net:     u.Scheme,
+			Timeout: timeout,
+		}
+
+		// If u.Host is IPv6 then wrap it in brackets
+		if strings.Count(u.Host, ":") >= 2 && !strings.HasPrefix(u.Host, "[") && !strings.Contains(u.Host, "]") {
+			u.Host = fmt.Sprintf("[%s]", u.Host)
+		}
+
+		address := u.Host
+		if u.Port() == "" {
+			address += fmt.Sprintf(":%d", h.dnsPort)
+		}
+
+		var resp, _, err = client.Exchange(msg, address)
+		if err != nil {
+			responseCh <- nil
+			return
+		}
+
+		responseCh <- resp
+	}(&dnsIP, newMsg.Copy())
+
+	var resp = h.waitResponse(ctx, responseCh)
+
+	if resp == nil {
+		dns.HandleFailed(rw, newMsg)
+		return
+	}
+
+	if err := rw.WriteMsg(resp); err != nil {
+		dns.HandleFailed(rw, newMsg)
+		return
+	}
+}
+
+func (h *dnsHealthCheckHandler) waitResponse(ctx context.Context, respCh <-chan *dns.Msg) *dns.Msg {
+	var respCount = cap(respCh)
+	for {
+		select {
+		case resp, ok := <-respCh:
+			if !ok {
+				return nil
+			}
+			respCount--
+			if resp == nil {
+				if respCount == 0 {
+					return nil
+				}
+				continue
+			}
+			if resp.Rcode == dns.RcodeSuccess {
+				return resp
+			}
+			if respCount == 0 {
+				return nil
+			}
+
+		case <-ctx.Done():
+			return nil
+		}
+	}
+}
+
+// NewTestDNSHandler creates a new health check dns handler
+// dnsHealthCheckHandler is expected to be placed at the beginning of the handlers chain
+func NewTestDNSHandler(params HealthCheckParams, opts ...TestDNSOption) dnsutils.Handler {
+	var h = &dnsHealthCheckHandler{
+		dnsServerIP: params.DNSServerIP,
+		healthHost:  ToWildcardPath(params.HealthHost),
+		scheme:      params.Scheme,
+		dnsPort:     53,
+	}
+
+	for _, o := range opts {
+		o(h)
+	}
+	return h
+}
+
+// ToWildcardPath will modify host by adding dot at the end
+func ToWildcardPath(host string) string {
+	return fmt.Sprintf("%s.", host)
+}
+
+// Option modifies default health check dns handler values
+type TestDNSOption func(*dnsHealthCheckHandler)
+
+// WithDefaultDNSPort sets default DNS port for health check dns handler if it is not presented in the client's URL
+func WithDefaultDNSPort(port uint16) TestDNSOption {
+	return func(h *dnsHealthCheckHandler) {
+		h.dnsPort = port
+	}
 }
